requirements.txt
keras
matplotlib
numpy
pandas
Pillow
scikit-learn
yfinance
streamlit
tensorflow